Added an Abstract class Command
The abstract class Command is created to better apply the command
pattern in the project; it contains a set of very general methods and
variables which is subject to change according to respective
subclasses. As new types of command are added to the project, the class
could be extended to have suitable methods and variables, such as
different types of content.