Added validation method for commandContents
For some commands(such as add and delete), it is important to ensure
that the content of the command line is as designated. Therefore, a
validation method is implemented to test out if the contents are
compliant to the standard before making the contents into a Command
object. Right now it just support “add”,  “delete” and “mark”. Will
extend in the future.